+++ CHANGES.md
@@ -35,6 +35,7 @@ Changes in 11.0
 - Use DBUpdate() & DBInsert() functions, program wide
 - Create missing Food Service Student Account in Accounts.php
 - Fix PHP warning if Food Service Student Account missing in Student.inc.php, Accounts.php, ServeMenus.php & Transactions.php
+- Let user edit inputs on Food Service User Account creation in Accounts.php
 
 Changes in 10.9.6
 -----------------